+++ data/scripts/constellations.py
@@ -4,12 +4,19 @@ import pandas as pd
 from shapely.geometry import Polygon
 
 from starplot.data import constellations
+from starplot.data.translations import language_name_column
 
 from data_settings import BUILD_PATH, RAW_PATH
+from translations import get_translations
 
 DATA_PATH = RAW_PATH / "iau"
 CRS = "+ellps=sphere +f=0 +proj=latlong +axis=wnu +a=6378137 +no_defs"
 
+translated = get_translations("constellation_names.csv")
+language_columns = [
+    (language, language_name_column(language))
+    for language in translated.keys()
+]
 
 def parse_ra(ra_str):
     """Parses RA from border file HH MM SS to 0...360 degree float"""
@@ -86,6 +93,10 @@ def build():
     constellation_records = build_constellations()
     df = pd.DataFrame.from_records(constellation_records)
 
+    for language, column_name in language_columns:
+        df[column_name] = df.apply(lambda d: translated[language].get(d.iau_id), axis=1)
+
+    print(df)
     gdf = gpd.GeoDataFrame(
         df,
         geometry=df["geometry"],